Add goreleaser (#29)
* WIP setting up goreleaser

* Some make refactoring

* Updated CI workflow configuration

Renamed the CI configuration file for better consistency. Also, replaced a manual run command with the use of goreleaser-action in the workflow. This change simplifies the setup process and ensures we're always using the latest version of goreleaser.

* Updated clean-up process in Makefiles

The clean-up process in the Makefiles has been updated. The changes include:
- Removed echo statements that were used for logging.
- Adjusted the depth of directories to be cleaned.
- Added a list command after deletion to provide feedback on what was removed.

* Updated protoc-gen-go version

The protoc-gen-go version has been updated from v1.33.0 to v1.34.0 in the generated code comments. This change reflects the use of a newer version of the tool for generating protocol buffer code.

* Refined clean-up process in Makefile

The clean-up process in the Makefile has been improved. The 'clean' command now includes additional targets for more thorough cleaning. A new 'clean_src' target was added to handle source file cleanup, and a 'clean_cli' target was introduced for CLI-related cleanup tasks. In addition, the 'clean' command in the cli/Makefile has been expanded to include uml2go and uml2pcl directories.

* Added GoReleaser to GitHub workflow

The GitHub workflow has been updated to include a new job for GoReleaser. This job runs on push events and depends on the build-go job. It sets up Go with version 1.22.2 or later, then runs GoReleaser with the latest version and clean args.

Changes have also been made to .goreleaser.yaml file, removing hooks and modifying builds section. The release section is now set to draft mode and auto pre-release.

A new 'release' target has been added in Makefile which triggers goreleaser with snapshot and clean options.

* Updated CI workflow

Added a step in the continuous integration workflow to checkout code using actions/checkout@v4. This will ensure that the latest version of the code is always used during the CI process.