Refactor NewGenCmd function
The NewGenCmd function has been refactored to use a new GeneratorOptions struct instead of the previous uml.Generator interface. This change allows for more flexibility in specifying generator options, such as the target. The creation and usage of the generator have also been moved inside the RunE function for better encapsulation.