Add abstractions project (#44)
* Delete UnitTest1 and SerilogConsole classes

- Remove UnitTest1 class from the test project
- Delete SerilogConsole class and related code from the main project

* Update TokenBinder to use readonly for Value field

The commit updates the TokenBinder class to use 'readonly' for the 'Value' field instead of initializing it directly.

* Added new Abstractions project

A new project named 'Abstractions' has been added to the solution. This includes setting up the necessary configurations for both Debug and Release modes. The project targets .NET 9.0 framework and includes a package reference to JetBrains.Annotations version 2023.3.0.

* Enabled implicit usings and nullable reference types

In this update, we've enabled implicit usings and nullable reference types in the build properties. We've also introduced new interfaces: IConverter, IConverterFrom, IConverterTo, and IGenerator. These interfaces provide methods for asynchronous conversion and generation tasks with cancellation support.

* Introduced IRunner interface

A new interface, IRunner, has been added to the Abstractions namespace. This interface inherits from both IConverter and IGenerator.

* Make tidy