Added FSharp test project and Docker tests
A new FSharp test project has been added to the solution. This includes a new DockerTests.fs file, which contains tests for performing IO with a container. A Program.fs file was also created as an entry point for the test project. The solution file has been updated to include this new project and its configurations.