Updated dependencies to newer versions
The commit includes an update of the 'unstoppablemango/tdl' package and its 'gen' module from version v0.0.3 to v0.0.4 in both cli and pkg modules. This change ensures that we are using the latest version of our dependencies, potentially benefiting from bug fixes, performance improvements or new features introduced in these updated versions.