Remove replacements from go modules (#62)
* Remove replacements

I think we'll need to create tags like pkg/v0.0.3 to get this working properly. And it's probably better to have goreleaser do that for us instead of creating 4 tags every time we want a release.

* Removed 'cli' from workflow matrix

The 'cli' module was removed from the GitHub Actions workflow matrix. This change simplifies the testing process by focusing on the 'gen' and 'pkg' modules only.

* Updated dependencies to newer versions

The commit includes an update of the 'unstoppablemango/tdl' package and its 'gen' module from version v0.0.3 to v0.0.4 in both cli and pkg modules. This change ensures that we are using the latest version of our dependencies, potentially benefiting from bug fixes, performance improvements or new features introduced in these updated versions.