Updated protoc-gen-go version and enhanced protobuf documentation
The protoc-gen-go version has been updated from v1.34.0 to v1.34.1, ensuring we're using the latest stable release for our protocol buffers.

In addition, improvements have been made to the protobuf documentation in 'any.ts'. This includes an additional example of unpacking a message and updates to existing examples for better clarity. A note about the lack of widely used type server implementations as of May 2023 was also added.