Its actually working now (#67)
* It's fuckin there bro

* Refactor EnsureBroker and introduce Patterns class

The EnsureBroker class has been refactored to separate concerns. A new static partial class, Patterns, has been introduced to hold the ApplicationStarted Regex pattern. The visibility of this method has also been changed from private to public. In the EnsureBroker class, a static readonly field for ApplicationStarted is now initialized with the Pattern's ApplicationStarted method. This change simplifies the usage of this pattern in other parts of the codebase by making it more accessible and reusable.

* GET IT BOI

* Enhanced logging and configuration handling

- Introduced dynamic console log level based on configuration
- Simplified command line builder syntax
- Added environment variable checks for debug and verbose modes in Config.fs
- Removed unused imports from Config.fs

* Updated protoc-gen-go version and enhanced protobuf documentation

The protoc-gen-go version has been updated from v1.34.0 to v1.34.1, ensuring we're using the latest stable release for our protocol buffers.

In addition, improvements have been made to the protobuf documentation in 'any.ts'. This includes an additional example of unpacking a message and updates to existing examples for better clarity. A note about the lack of widely used type server implementations as of May 2023 was also added.

* Updated .gitignore and Makefile

Added .envrc to the .gitignore file. In the Makefile, a new 'dev' target has been introduced for setting up the development environment. Also, added a rule to regenerate .envrc when necessary. Enhanced linting process by including a step to touch '.make/regen_envrc'.

* Updated Makefile for dev environment management

Added a new 'undev' target to the Makefile that tears down the development environment. This includes removing .envrc and .make/regen_envrc files. The 'dev' target now ensures the setup of the development environment.

* Updated environment handling and config structure

Significant changes include the addition of an 'env' variable to handle different environments based on whether it's a debug or release build. The Config type has been refactored to use static members instead of instance members, which now reference the new 'env' variable for more consistent configuration across different environments.

* Added retry logic to socket connection

Introduced a maximum retry limit for establishing the socket connection. If the connection is not established within 10 attempts, an exception is thrown. Also, replaced usage of local config object with global Config object in creating client channel.

* Fix formatting