Updated Makefile with versioning and echo command
The Makefile has been updated to include a new versioning feature. A VERSION variable has been added that uses the 'dotnet minver' command to get the current version of the project. This version is then exported as MINVERVERSIONOVERRIDE. Additionally, a new 'version' target has been added which echoes out the current project version when called.