Added Config class in Broker.Internal
Introduced a new static class, Config, within the Broker.Internal namespace. This class includes a private method to check if the application is running in a Docker container and public constants for GitHub product name and plugin directory path. The plugin directory path varies based on whether the app is running in Docker or not.