Refactor Broker Program setup
The Broker Program setup has been refactored to improve the handling of GitHub client and plugin services. The GitHub product string is now fetched from a configuration file instead of being hardcoded. Additionally, an HttpClient for GitHub has been added to the service collection. A new temporary subdirectory is created for the broker socket in development mode, replacing the previous method of using a temp path. Lastly, PluginService has been added as a hosted service.