Fetch plugins on broker start (#73)
* Remove unnecessary dep from CLI

* Mount plugin dir as tmpfs

* Updated package references

The commit includes updates to the package references in the project file. The version of "Microsoft.Extensions.DependencyInjection.Abstractions" has been updated to 8.0.1, and "Serilog.Extensions.Hosting" has been removed. No changes were made to the project references.

* Removed unused imports

Two unnecessary using directives were removed from the DockerBroker and Program files. This cleanup helps to maintain code readability and efficiency.

* Added Octokit package to Broker project

The Broker project now includes the Octokit package, version 11.0.1. This addition will enhance GitHub API interactions within the application.

* Added GitHub integration to Broker service

The Broker service now includes a transient GitHub client, allowing it to interact with the GitHub API. This was achieved by adding Octokit as a dependency and initializing a new instance of the GitHubClient class in the service configuration. The socket path has also been slightly adjusted for better readability.

* Added PluginService for plugin management

This update introduces a new service, PluginService, which is responsible for managing plugins. It fetches the latest plugin release from GitHub and checks if it matches the current system's architecture and OS. If there's a match, it retrieves the download URL of the asset. The service also includes methods to determine the system's architecture and OS type. Error logging has been implemented in case of failure at any step.

* Updated .editorconfig settings

Added a new configuration line to the ReSharper properties in the .editorconfig file. This change disables indentation for switch labels in C#.

* Added Config class in Broker.Internal

Introduced a new static class, Config, within the Broker.Internal namespace. This class includes a private method to check if the application is running in a Docker container and public constants for GitHub product name and plugin directory path. The plugin directory path varies based on whether the app is running in Docker or not.

* Added HttpClients static class

A new static class named HttpClients has been introduced in the Broker.Internal namespace. This class currently contains a single constant string for GitHub.

* Refactor Broker Program setup

The Broker Program setup has been refactored to improve the handling of GitHub client and plugin services. The GitHub product string is now fetched from a configuration file instead of being hardcoded. Additionally, an HttpClient for GitHub has been added to the service collection. A new temporary subdirectory is created for the broker socket in development mode, replacing the previous method of using a temp path. Lastly, PluginService has been added as a hosted service.

* Enhanced PluginService with archive handling

Significant changes have been made to the PluginService. The service now includes additional dependencies for HTTP client factory and logger. It also checks if a plugin directory exists, creates one if it doesn't, and handles exceptions during creation. The code fetches the latest GitHub release only when an asset file does not exist locally. It then downloads the asset from GitHub and writes it to a local file.

The service can now handle both tar.gz and zip archives. For tar.gz files, it uses GZipStream for decompression before extracting contents using TarFile.ExtractToDirectoryAsync(). For zip files, it uses ZipFile.ExtractToDirectory() directly.

Minor changes include updating TryGetArchitecture() to return true when architecture is not null or whitespace, instead of the opposite.