Implement the upgrade command (#74)
* Enhanced broker upgrade command

The broker upgrade command has been enhanced with new options. Users can now specify whether they want to upgrade to the latest version or a specific version. The code handling these upgrades has also been updated accordingly. Additionally, default values have been added for CancellationToken parameters in IBroker interface methods.

* Refactored DockerBroker class

The DockerBroker class has been refactored for better readability and maintainability. The changes include:
- Replaced the IDictionary with a Dictionary for Labels.
- Extracted the creation of default start arguments into a separate method, DefaultStartArgs().
- Simplified the process of starting a broker by using DefaultStartArgs() method.
- Created a new Find() method to find matching containers, which is now used in Status() and Stop() methods.
- Implemented Upgrade() method to handle broker upgrades. This includes checking if an upgrade is necessary, stopping the current broker if it exists, and starting a new one with the upgraded version.

* Updated DockerBroker and Handlers logic

Enhanced the version handling in DockerBroker and Handlers. In DockerBroker, added a check to set version as 'latest' if it's null or '0.0.0'. In Handlers, modified the assembly version retrieval process to return null when it fails and formatted the returned version string for successful cases.

* Updated Docker tag handling in dev mode

The Docker tag handling has been updated to better support development environments. Now, the 'local' tag is used when running in a development environment and 'latest' is used otherwise. This change also includes an override for the version in development mode to use the 'local' tag. The requirement for a Tag property in StartArgs has been removed, allowing it to be optional. The containerTag function from Config.fs was also removed as it's no longer needed with these changes.