Add a sort of testing framework (#122)
* Added new projects to solution

Two new projects, 'actions' and 'build-cs', have been added to the solution. The corresponding project sections and their items have also been included. Additionally, these new projects have been linked with existing ones in the GlobalSection.

* Enable aot again

* Updated project SDK and added new package

The project's SDK has been updated. Additionally, a new package reference, Docker.DotNet with version 3.125.15, has been included in the FSharp project file.

* Added Docker functionality to FSharp project

A new module, UnMango.Tdl.Docker, has been added to the FSharp project. This module includes a Progress type and a Workload type with associated functions for creating and managing workloads. A private Builder type is also introduced which provides methods for interacting with Docker containers such as creating images, attaching to containers, starting/stopping containers and cleaning up resources. The Docker.fs file has been included in the FSharp project file.

* Added FSharp test project and Docker tests

A new FSharp test project has been added to the solution. This includes a new DockerTests.fs file, which contains tests for performing IO with a container. A Program.fs file was also created as an entry point for the test project. The solution file has been updated to include this new project and its configurations.

* Added Entrypoint to Workload and Builder

The Workload type now includes an 'Entrypoint' field. This is also initialized in the 'create' function within the Workload module. In the Builder module, both Cmd and Entrypoint fields are converted to ResizeArray before being passed to CreateContainerParameters.

* Updated Docker tests for better validation

Added Xunit to the DockerTests file and improved the test case 'Can perform IO with a container'. Now, it not only checks if the output equals the message but also asserts that the output length is greater than zero. This ensures more robust testing by validating both content and existence of output.

Refactor Docker container handling in FSharp

The Docker container handling logic in the FSharp module has been refactored for better readability and efficiency. The changes include:
- Reordering of operations to ensure proper cleanup even when a container fails to start.
- Introduction of a new 'wait' function that waits for a container to finish its execution before proceeding.
- Streamlining the use of Async tasks and cancellation tokens.
- Removal of unnecessary try-finally block, as cleanup is now handled more systematically.

* Remove missing file from project

WIP fiddling with arbs

Set labels arb

More generators

More arbs

Added RunnerTest module in Testing

A new testing module, RunnerTest, has been added to the project. This module includes a function for round-trip testing of Tdl.Gen and Tdl.From functions. The fsproj file has also been updated to include this new test file.

Adding the test runner

Enhanced binary existence check and testing

The code changes have improved the binary existence verification process by adding a null check. In case of a missing binary, an appropriate message is now displayed. The testing mechanism has also been enhanced with the introduction of 'gen' and 'from' variables for roundTrip function in RunnerTest.

Use the arb for the test

Clean

Refactor main

Refactor main, except it compiles this time

Better arg validation

* Testing tests lol

Introduce a stack overflow

* Updated Spec.fs to debug stack overflow

Commented out the generation and addition of labels, meta, and types in Spec.fs. This is a temporary measure to isolate the cause of a recurring stack overflow issue. Further investigation is needed to identify the root cause and implement a permanent fix.

* Updated test functions and added new ones

Enhanced the 'roundTrip' function in the testing module to use 'Equals' method for comparison instead of equality operator. Added two new async functions, 'generateData' and 'consumeData', to respectively generate data from a given spec into a stream and consume data from a stream into a spec, checking if the consumed spec equals the original one.

* Renamed test module and added new tests

The test module has been renamed for clarity. Additionally, two new property-based tests have been introduced to verify data generation and consumption. The project file has also been updated to reflect the renaming of the test file.

* Updated error message for missing path

The error message displayed when no path is provided has been updated to be more user-friendly and clear. Instead of "Path was not provided", it now says "<path> is required".

* Updated Docker test property attributes

The Docker test property attributes have been updated. The 'Skip' attribute has been added to the Property decorator in the DockerTests file, indicating that a specific test is currently not functioning as expected and should be skipped during testing.