Updated test configuration based on environment
The number of maximum tests to run is now determined by the presence of a "CI" environment variable. If the variable is not set or empty, 100 tests will be run; otherwise, it will run 500 tests. This change allows for more extensive testing in CI environments while keeping local testing relatively quick.