Refactor CLI constructor with options pattern
The CLI constructor has been refactored to use the options pattern. This allows for more flexible configuration of the CLI instance. A new function `WithArgs` has been added to append arguments to a CLI instance. The changes improve code readability and maintainability by making it easier to add new configuration options in the future.