End to end test for the main cli (#202)
* Go crap


Go crap


Go crap

* Added plugin support to CLI commands

The 'from' and 'gen' CLI commands have been updated to include plugin support. The changes involve getting the plugin name, looking up the plugin, and executing it with a runner. Error handling has also been added for each step of this process.

* Improved error logging and context usage

Enhanced the error logging mechanism to provide more detailed information when a test fails. Also, replaced the use of context.TODO() with context.Background() for better clarity and correctness in the codebase.

* Refactor test runner and update protobuf handling

Significant changes include:
- Removed the context import and replaced the runner package with os/exec for command execution.
- Updated how protobufs are handled by marshalling the spec into data before running the command.
- Simplified error handling by removing redundant checks.
- Streamlined reading of actual output from buffer, reducing complexity.

* Refactor error handling in test runner

The test runner's error handling has been improved. The unmarshaling process now initializes a new Spec object directly, rather than using a pointer. Additionally, the command execution now captures stderr and includes it in the returned error if the command fails. This should provide more context when debugging failed tests.

* Guess what, go crap

* Updated output formatting for test results

The format of the printed output for test results has been modified. The 'Expected' and 'Actual' values are now printed on new lines, improving readability when dealing with multi-line strings.

* Updated Makefile for e2e testing

The Makefile has been updated to include 'bin/uml2ts' in the end-to-end (e2e) testing process. This change ensures that the UML to TypeScript conversion is tested as part of the e2e tests.

* Fixing the issue

* Add the other test case