Introduced runnerCmd in CLI
A new struct, runnerCmd, has been added to the command-line interface. This struct includes a method that runs an input function on either stdin or file arguments. It also logs debug information about the process such as whether data was found in stdin, if files were found in arguments, and whether these files could be opened successfully.