Added mediaType struct and methods
Introduced a new 'mediaType' struct in the runner package, along with associated methods. This includes a 'WithMediaType' function for creating instances of this type. Also added two interfaces, 'From' and 'Gen', to the uml package's converter.go and generator.go files respectively. These changes are part of an effort to enhance modularity and code organization.