Refactor generic generator interface and functions
The generic generator interface has been refactored for better clarity and flexibility. The previous single function interface has been replaced with a new Generator interface that includes separate input() and output() methods. Additionally, two mapping functions have been added to allow transformations on both the input and output of a Generator. A Run function is also introduced to execute the generator with given input and output.